Dynamically Typed Indexes

+++ src/storage/file_io_tns.cpp
@@ -85,7 +85,7 @@ void writeTNS(std::string filename, const TensorBase& tensor) {
 
 void writeTNS(std::ostream& stream, const TensorBase& tensor) {
   for (auto& value : iterate<double>(tensor)) {
-    for (int coord : value.first) {
+    for (size_t coord : value.first) {
       stream << coord+1 << " ";
     }
     stream << value.second << endl;
